powerpc/inst: Avoid undefined behaviour in ppc_inst_write()
In C, it is undefined behavior to cast and store to an address as a u64
pointer if that address is not u64 aligned. Prefixed instructions are
not necessarily u64 aligned so by performing such a cast and store,
ppc_inst_write() is relying on the compiler implementation to generate
"correct" code.

The prefix instructions should be written with a single store
instruction as a double word to avoid the situation of having a partial
instruction in the kernel image.

To avoid relying on undefined behavior use inline assembly for this
store.

Fixes: 309a0a601864 ("powerpc/code-patching: Replace patch_instruction() by ppc_inst_write() in selftests")
Signed-off-by: Jordan Niethe <jniethe5@gmail.com>